+++ src/___tests___/components/ng-redux.spec.ts
@@ -2,9 +2,6 @@ import 'reflect-metadata';
 import {expect, use} from 'chai';
 import { createStore } from 'redux';
 import {NgRedux} from '../../components/ng-redux';
-import {dispatch} from '../../components/dispatch';
-import {dispatchAll} from '../../components/dispatchAll';
-import {select} from '../../components/select';
 import * as sinon from 'sinon';
 import * as sinonChai from 'sinon-chai';
 import * as _ from 'lodash';
@@ -292,171 +289,3 @@ describe('NgRedux Observable Store', () => {
 
 });
 
-describe('Decorator Interface', () => {
-
-  function mockActionCreator() {
-    return {
-      type: 'INCREMENT_COUNTER'
-    };
-  }
-
-  describe('@dispatch', () => {
-
-    let store;
-    let connector;
-    let targetObj;
-    let defaultState;
-
-    beforeEach(() => {
-      defaultState = {
-        foo: 'bar',
-        baz: -1
-      };
-      store = createStore((state = defaultState, action) => {
-        const newState = Object.assign({}, state, { baz: action.payload });
-        return newState;
-      });
-      targetObj = {};
-      connector = new NgRedux(store);
-    });
-
-
-    it('attaches an instance method on the decorated property', () => {
-
-      class MockClass {
-        @dispatch(mockActionCreator) anInstanceMethod;
-      }
-      const mockInstance = new MockClass();
-      expect(mockInstance.anInstanceMethod).to.be.instanceOf(Function);
-
-    });
-
-  });
-
-  describe('@dispatchAll', () => {
-
-    const actionCreatorsObject = {
-      mockActionCreator: mockActionCreator,
-      anotherFunction: () => ({ type: 'ANOTHER_ACTION' }),
-      thisShouldNotBeBound: 'I`m a string not a function'
-    };
-
-    @dispatchAll(actionCreatorsObject)
-    class MockClass {
-      mockActionCreator;
-      anotherFunction;
-      thisShouldNotBeBound; // typescripted
-    }
-
-    let mockInstance = new MockClass();
-
-    it('attaches all functions on the given object, to the decorated class',
-       () => {
-      expect(mockInstance.mockActionCreator).to.be.instanceOf(Function);
-      expect(mockInstance.anotherFunction).to.be.instanceOf(Function);
-    });
-
-    it('only attaches functions, ignoring anything else on the object', () => {
-      expect(mockInstance.thisShouldNotBeBound).to.be.undefined;
-    });
-
-    it('wraps the action creator in a dispatch call', () => {
-      expect(mockInstance.mockActionCreator)
-        .to.not.equal(actionCreatorsObject.mockActionCreator);
-      expect(mockInstance.anotherFunction)
-        .to.not.equal(actionCreatorsObject.anotherFunction);
-    });
-
-  });
-
-  describe('@select', () => {
-
-    let ngRedux;
-    let targetObj;
-    let defaultState;
-    let rootReducer;
-    let mockAppRef;
-  
-    beforeEach(() => {
-      defaultState = {
-        foo: 'bar',
-        baz: -1
-      };
-      rootReducer = (state = defaultState, action) => {
-        const newState = Object.assign({}, state, { baz: action.payload });
-        return newState;
-      };
-      targetObj = {};
-      mockAppRef = {
-        tick: sinon.spy()
-      };
-      ngRedux = new NgRedux(mockAppRef);
-      ngRedux.configureStore(rootReducer, defaultState);
-    });
-
-    describe('when left empty', () => {
-
-      it('automatically attempts to bind to a store property that matches the' +
-         ' name of the class property', () => {
-
-        class MockClass {
-          @select() foo: any;
-        }
-
-        let mockInstance = new MockClass();
-
-        expect(mockInstance.foo.subscribe).to.be.instanceOf(Function);
-
-      });
-
-      it('attempts to bind by name ignoring any $ characters in the class ' +
-         'property name', () => {
-
-         class MockClass {
-          @select() foo$: any;
-        }
-
-        let mockInstance = new MockClass();
-
-        expect(mockInstance.foo$.subscribe).to.be.instanceOf(Function);
-
-      });
-
-    });
-
-    describe('when passed a string', () => {
-
-      it('attempts to bind to the store property whose name matches the ' +
-         'string value', () => {
-
-        class MockClass {
-          @select('foo') asdf: any;
-        }
-
-        let mockInstance = new MockClass();
-
-        expect(mockInstance.asdf.subscribe).to.be.instanceOf(Function);
-
-      });
-
-    });
-
-    describe('when passed a function', () => {
-
-      it('attempts to use that function as the selector function', () => {
-
-        class MockClass {
-          @select(state => state.foo) asdf: any;
-        }
-
-        let mockInstance = new MockClass();
-
-        expect(mockInstance.asdf.subscribe).to.be.instanceOf(Function);
-
-      });
-
-    });
-
-  });
-
-});
