+++ src/decorators/index.ts
@@ -1,0 +1,3 @@
+export { select } from './select';
+export { dispatch } from './dispatch';
+export { dispatchAll } from './dispatch-all';
