+++ src/index.ts
@@ -1,4 +1,3 @@
 export { NgRedux } from './components/ng-redux';
-export { dispatch } from './components/dispatch';
-export { select } from './components/select';
-export { dispatchAll } from './components/dispatchAll';
+export { dispatch, select, dispatchAll } from './decorators';
+
